+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Implement CEILING(n, m=1)
 * Implement FLOOR(n, m=1)
 * Implement ROUND(n, m=0)
+* Implement TRUNC(n, m=0)
 
 
 ## xel.js 1.3.0  released 2024-04-05